# No more gd `bitvec`