## Motivation

No more gd `bitvec`.

```
Before
-------
Pedersen CRH setup      time:   [635.04 us 636.57 us 638.71 us]                               
Found 3 outliers among 50 measurements (6.00%)
  1 (2.00%) high mild
  2 (4.00%) high severe

Pedersen CRH hash       time:   [62.865 us 62.992 us 63.120 us]                             
Found 4 outliers among 50 measurements (8.00%)
  3 (6.00%) high mild
  1 (2.00%) high severe

After
------
Pedersen CRH setup      time:   [636.88 us 639.05 us 642.13 us]                               
                        change: [-0.2128% +0.5658% +1.5633%] (p = 0.24 > 0.05)
                        No change in performance detected.
Found 2 outliers among 50 measurements (4.00%)
  1 (2.00%) high mild
  1 (2.00%) high severe

Pedersen CRH hash       time:   [63.615 us 63.723 us 63.845 us]                             
                        change: [+0.5504% +0.8897% +1.2171%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 3 outliers among 50 measurements (6.00%)
  1 (2.00%) high mild
  2 (4.00%) high severe
```